Fix password hashing in get_sha256_hash_from_str
Add 1 to the len_strs_plus_salt when passing it as the maxlen to the snprintf-function to account for the terminating null character. Previously the last character of the password was truncated.

Change the type of the hash_raw variable from char to unsigned char because mhash_init expects an unsigned char. Previously this would result in a wrong hexadecimal representation of the hash on platforms where the char type is a signed char. Bytes from 00 to 7f were correctly represented, but bytes from 80 to ff were always represented as ff in the hash.