Fix input buffer sanitization
This reverts a change from commit dac2ee4f causing some inputs to be processed incorrectly when using a mysql database. \r would not be removed anymore for inputs ending with \r\n because the code would only remove the \n leaving the \r in place. The function mysql_real_escape_string would then replace the \r with \\r causing some nntp commands to be processed incorrectly. For example the input "group x\r\n" would be transformed to "group x\r" which would be transformed to "group x\\r". This would result in searching the group with the name "x\\r" instead of "x".